Type of LSN#position was decrease from long to int.

+++ core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/paginated/wal/OMemoryWriteAheadLog.java
@@ -27,6 +27,7 @@ import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
 /**
@@ -34,7 +35,7 @@ import java.util.concurrent.atomic.AtomicLong;
  * @since 6/25/14
  */
 public class OMemoryWriteAheadLog extends OAbstractWriteAheadLog {
-  private final AtomicLong nextPosition = new AtomicLong();
+  private final AtomicInteger nextPosition = new AtomicInteger();
 
   @Override
   public OLogSequenceNumber begin() {
